+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeListPage/CodeListsActionsBar/AddCodeListDropdown/index.ts
@@ -1,0 +1,1 @@
+export { AddCodeListDropdown } from './AddCodeListDropdown';
