+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeListPage/CodeListsActionsBar/CodeListsActionsBar.tsx
@@ -1,16 +1,11 @@
-import React, { useRef } from 'react';
+import React from 'react';
 import type { TextResource } from '@studio/components-legacy';
 import { StudioSearch } from '@studio/components-legacy';
 import type { ChangeEvent } from 'react';
 import classes from './CodeListsActionsBar.module.css';
 import { useTranslation } from 'react-i18next';
 import type { CodeListWithMetadata } from '../CodeListPage';
-import { CreateNewCodeListModal } from './CreateNewCodeListModal';
-import { FileNameUtils } from '@studio/pure-functions';
-import { useUploadCodeListNameErrorMessage } from '../hooks/useUploadCodeListNameErrorMessage';
-import { toast } from 'react-toastify';
-import { StudioDropdown } from '@studio/components';
-import { PlusCircleIcon, PlusIcon, UploadIcon } from '@studio/icons';
+import { AddCodeListDropdown } from './AddCodeListDropdown';
 
 export type CodeListsActionsBarProps = {
   onBlurTextResource?: (textResource: TextResource) => void;
@@ -30,29 +25,11 @@ export function CodeListsActionsBar({
   textResources,
 }: CodeListsActionsBarProps) {
   const { t } = useTranslation();
-  const addCodeListRef = useRef<HTMLDialogElement>(null);
-
-  const getInvalidUploadFileNameErrorMessage = useUploadCodeListNameErrorMessage();
 
   const handleChangeSearch = (event: ChangeEvent<HTMLInputElement>) =>
     onSetSearchString(event.target.value);
 
   const handleClearSearch = () => onSetSearchString('');
-
-  const onSubmit = (file: File) => {
-    const fileNameError = FileNameUtils.findFileNameError(
-      FileNameUtils.removeExtension(file.name),
-      codeListNames,
-    );
-    if (fileNameError) {
-      return toast.error(getInvalidUploadFileNameErrorMessage(fileNameError));
-    }
-    onUploadCodeList(file);
-  };
-
-  const handleOpenAddCodeListDialog = () => {
-    addCodeListRef.current?.showModal();
-  };
 
   return (
     <div className={classes.actionsBar}>
@@ -62,32 +39,12 @@ export function CodeListsActionsBar({
         clearButtonLabel={t('app_content_library.code_lists.clear_search_button_label')}
         onClear={handleClearSearch}
       />
-      <StudioDropdown
-        triggerButtonVariant='secondary'
-        triggerButtonText={t('app_content_library.code_lists.add_new_code_list')}
-        icon={<PlusIcon />}
-      >
-        <StudioDropdown.Item>
-          <StudioDropdown.Button onClick={handleOpenAddCodeListDialog} icon={<PlusCircleIcon />}>
-            {t('app_content_library.code_lists.create_new_code_list')}
-          </StudioDropdown.Button>
-        </StudioDropdown.Item>
-        <StudioDropdown.Item>
-          <StudioDropdown.FileUploaderButton
-            icon={<UploadIcon />}
-            onFileUpload={onSubmit}
-            fileInputProps={{ accept: '.json' }}
-          >
-            {t('app_content_library.code_lists.upload_code_list')}
-          </StudioDropdown.FileUploaderButton>
-        </StudioDropdown.Item>
-      </StudioDropdown>
-      <CreateNewCodeListModal
+      <AddCodeListDropdown
         codeListNames={codeListNames}
         onBlurTextResource={onBlurTextResource}
+        onUploadCodeList={onUploadCodeList}
         onUpdateCodeList={onUpdateCodeList}
         textResources={textResources}
-        ref={addCodeListRef}
       />
     </div>
   );
